Fix some warnings

+++ MahApps.Metro/Controls/MetroWindow.cs
@@ -942,10 +942,10 @@ namespace MahApps.Metro.Controls
                     IntPtr windowHandle = new WindowInteropHelper(this).Handle;
                     UnsafeNativeMethods.ReleaseCapture();
                     var wpfPoint = this.PointToScreen(mPoint);
-                    var x = Convert.ToInt16(wpfPoint.X);
-                    var y = Convert.ToInt16(wpfPoint.Y);
-                    var lParam = (int) (uint) x | (y << 16);
-                    UnsafeNativeMethods.SendMessage(windowHandle, Constants.WM_NCLBUTTONDOWN, Constants.HT_CAPTION, lParam);
+                    var x = (uint)Convert.ToInt16(wpfPoint.X);
+                    var y = (uint)Convert.ToInt16(wpfPoint.Y);
+                    var lParam =   x | (y << 16);
+                    UnsafeNativeMethods.SendMessage(windowHandle, Constants.WM_NCLBUTTONDOWN, Constants.HT_CAPTION, (int)lParam);
                 }
 
                 var canResize = this.ResizeMode == ResizeMode.CanResizeWithGrip || this.ResizeMode == ResizeMode.CanResize;
